feat(login regster): add login and register functions
login and register api functions in try and catch  block that send psot request to api and display an error message in ui. register logs in after success, login redirects to home after success.

onlogin and onregister bind the api functions to the forms and buttons in the ui. take form data through class FormData. submit form with submit button. #2